move all files around to work with godep
