+++ src/Gopkg.toml
@@ -22,9 +22,21 @@
 
 
 [[constraint]]
+  name = "github.com/dgrijalva/jwt-go"
+  version = "3.2.0"
+
+[[constraint]]
   name = "github.com/go-ozzo/ozzo-routing"
   version = "2.1.4"
 
 [[constraint]]
-  branch = "master"
-  name = "github.com/tylergeery/trash_hunt"
+  name = "github.com/gomodule/redigo"
+  version = "2.0.0"
+
+[[constraint]]
+  name = "github.com/goware/emailx"
+  version = "0.2.0"
+
+[[constraint]]
+  name = "github.com/lib/pq"
+  version = "1.1.1"
