+++ src/api_server/controllers/test/auth_test.go
@@ -7,9 +7,9 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/tylergeery/trash_hunt/auth"
-	"github.com/tylergeery/trash_hunt/game"
-	"github.com/tylergeery/trash_hunt/test"
+	"github.com/tylergeery/trash_hunt/src/auth"
+	"github.com/tylergeery/trash_hunt/src/game"
+	"github.com/tylergeery/trash_hunt/src/test"
 )
 
 // TestCreateWithInvalidKey
