+++ src/api_server/controllers/test/player_test.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/tylergeery/trash_hunt/game"
-	"github.com/tylergeery/trash_hunt/test"
+	"github.com/tylergeery/trash_hunt/src/game"
+	"github.com/tylergeery/trash_hunt/src/test"
 )
 
 // TestPlayerCreateFailures
