+++ src/api_server/middleware/auth.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 
 	"github.com/go-ozzo/ozzo-routing"
-	"github.com/tylergeery/trash_hunt/auth"
+	"github.com/tylergeery/trash_hunt/src/auth"
 )
 
 type key string
