+++ src/auth/token_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	jwt "github.com/dgrijalva/jwt-go"
-	"github.com/tylergeery/trash_hunt/game"
+	"github.com/tylergeery/trash_hunt/src/game"
 )
 
 func TestCreatingAndExtractingToken(t *testing.T) {
