+++ src/game/maze_test.go
