+++ src/storage/persistent_test.go
@@ -4,7 +4,7 @@ import (
 	"database/sql"
 	"testing"
 
-	_ "github.com/tylergeery/trash_hunt/test"
+	_ "github.com/tylergeery/trash_hunt/src/test"
 )
 
 func TestInsertUpdateAndRemove(t *testing.T) {
