+++ src/test/test.go
@@ -1,27 +1,8 @@
 package test
 
 import (
-	"os"
-	"strings"
 	"testing"
 )
-
-// SetVars for testing
-func init() {
-	redis := os.Getenv("REDIS_HOST")
-	pg := os.Getenv("PG_HOST")
-
-	// set persistent storage
-	os.Setenv("DB_USER", "dev")
-	os.Setenv("DB_PASS", "dev_secret")
-	os.Setenv("DB_HOST", strings.Trim(string(pg), "\n'"))
-	os.Setenv("DB_TABLE", "dev_secret")
-	os.Setenv("DB_SSL_MODE", "disable")
-
-	// set temporary storage
-	os.Setenv("REDIS_HOST", strings.Trim(string(redis), "\n'"))
-	os.Setenv("REDIS_PORT", "6379")
-}
 
 // FatalOnError to be reused by testing package
 func FatalOnError(t *testing.T, err error) {
