--- tcp_server/Gopkg.lock
@@ -1,33 +1,0 @@
-# This file is autogenerated, do not edit; changes may be undone by the next 'dep ensure'.
-
-
-[[projects]]
-  name = "github.com/gomodule/redigo"
-  packages = ["internal","redis"]
-  revision = "9c11da706d9b7902c6da69c592f75637793fe121"
-  version = "v2.0.0"
-
-[[projects]]
-  name = "github.com/goware/emailx"
-  packages = ["."]
-  revision = "80a00e4677ff2f102e17b78107f255253e2032e5"
-  version = "v0.2.0"
-
-[[projects]]
-  name = "github.com/lib/pq"
-  packages = [".","oid","scram"]
-  revision = "bc6a3c0594130b1e34005880bc600b6d3f49fa7f"
-  version = "v1.1.1"
-
-[[projects]]
-  branch = "master"
-  name = "github.com/tylergeery/trash_hunt"
-  packages = ["game","storage"]
-  revision = "d562c51e40540c90050ed9f2c578e514d0ea6e2e"
-
-[solve-meta]
-  analyzer-name = "dep"
-  analyzer-version = 1
-  inputs-digest = "cc0c7415e5569bd88585c0c9dabc1816440262b010eb1e6b697db7fcbe3c5f8a"
-  solver-name = "gps-cdcl"
-  solver-version = 1
